Update esp-idf_component.rst
When I added Arduino using the previous proposed way, it did not appeared on the .gitmodules file. I think that cloning the repo is not enough to add a submodule, you have to specifically add it. Following the same steps as shown in documentation https://github.com/joltwallet/esp_littlefs I edited this doc.